+++ tests/dwd/test_index.py
@@ -9,7 +9,7 @@ from wetterdienst.dwd.index import (
 )
 from wetterdienst.dwd.metadata.constants import DWDCDCBase
 from wetterdienst.dwd.observations import (
-    DwdObservationParameterSet,
+    DwdObservationDataset,
     DwdObservationPeriod,
     DwdObservationResolution,
 )
@@ -18,7 +18,7 @@ from wetterdienst.util.network import list_remote_files
 
 def test_build_index_path():
     path = build_path_to_parameter(
-        DwdObservationParameterSet.CLIMATE_SUMMARY,
+        DwdObservationDataset.CLIMATE_SUMMARY,
         DwdObservationResolution.DAILY,
         DwdObservationPeriod.HISTORICAL,
     )
@@ -42,7 +42,7 @@ def test_list_files_of_climate_observations():
 def test_fileindex():
 
     file_index = _create_file_index_for_dwd_server(
-        DwdObservationParameterSet.CLIMATE_SUMMARY,
+        DwdObservationDataset.CLIMATE_SUMMARY,
         DwdObservationResolution.DAILY,
         DwdObservationPeriod.RECENT,
         DWDCDCBase.CLIMATE_OBSERVATIONS,